Fix: include kernel-azure in SLE12 SP5 On-Demand profiles
Due to the new keg feature that allows image parameters to be in base
profiles, common Azure modules could be moved into the base for reducing
duplication in the output. However, that means key 'azure-kernel' was now
included in 'Base-On-Demand' and thus cannot be overwritten in e.g.
Standard-On-Demand (to select 'kernel-azure') This commit fixes it by making
'kernel-default' an explicit include (only for SLE12 SP5).
Also cleanup SLE12 SP5 GCE profile (drop obsolete includes).