Use a subvolume for /var
Having var on a spare partition prevents the resizing of the other parts of the
filesyztem as those parts will be in the middle of the partition table. Since
we have an image size limit of 10GB in the Public Cloud using any space for
var and not being able to resize other parts of the file system creates a size
problem. Using a subvolum for var eliminates this problem.