Use cgroup v1 for CHOST flavor builds
SAP is reporting issues with the v2 cgroup setting and requesting we
go back to v1. Given that we have no othet known consumers of the CHOST
images we honor this request. That said the change to v1 may cause issues
with k8s (https://kubernetes.io/docs/concepts/architecture/cgroups/).