Base updates
There are few packages that directly depend on iptables, it was accidentally
dropped from one of our images when we dropped docker. iptables is considered
a fundamental tool and should always be in our images. Add it explicitly to
the base configuration. Rename the package for suseconnect. The Golang
implementation is available everywhere now.